Move ELF loader to separate translation unit
As we are going to expose 'memcpy' methods of FuncMemory, there is no
need to hide ELF loader inside the class, so it is a standalone function
now.
