+++ simulator/infra/memory/t/unit_test.cpp
@@ -9,6 +9,7 @@
 
 // MIPT-MIPS modules
 #include "../memory.h"
+#include "../elf/elf_loader.h"
 
 static const std::string valid_elf_file = TEST_DATA_PATH "mips_bin_exmpl.out";
 // the address of the ".data" section
@@ -29,8 +30,14 @@ TEST_CASE( "Func_memory_init: Process_Wrong_Args_Of_Constr")
 
 TEST_CASE( "Func_memory_init: Process_Correct_ElfInit")
 {
-    CHECK_NOTHROW( FuncMemory( ).load_elf_file( valid_elf_file));
-    CHECK_NOTHROW( FuncMemory( 48, 15, 10).load_elf_file( valid_elf_file));
+    FuncMemory mem;
+    CHECK_NOTHROW( ::load_elf_file( &mem, valid_elf_file));
+}
+
+TEST_CASE( "Func_memory_init: Process_Correct_ElfInit custom mapping")
+{
+    FuncMemory mem( 48, 15, 10);
+    CHECK_NOTHROW( ::load_elf_file( &mem, valid_elf_file));
 }
 
 TEST_CASE( "Func_memory_init: Process_Wrong_ElfInit")
@@ -38,13 +45,14 @@ TEST_CASE( "Func_memory_init: Process_Wrong_ElfInit")
     // test behavior when the file name does not exist
     const std::string wrong_file_name = "./1234567890/qwertyuiop";
     // must exit and return EXIT_FAILURE
-    CHECK_THROWS_AS( FuncMemory( ).load_elf_file( wrong_file_name), InvalidElfFile);
+    FuncMemory mem;
+    CHECK_THROWS_AS( load_elf_file( &mem, wrong_file_name), InvalidElfFile);
 }
 
 TEST_CASE( "Func_memory: StartPC_Method_Test")
 {
     FuncMemory func_mem;
-    func_mem.load_elf_file( valid_elf_file);
+    ::load_elf_file( &func_mem, valid_elf_file);
 
     CHECK( func_mem.startPC() == 0x4000b0u /*address of the ".text" section*/);
 }
@@ -52,7 +60,7 @@ TEST_CASE( "Func_memory: StartPC_Method_Test")
 TEST_CASE( "Func_memory: Read_Method_Test")
 {
     FuncMemory func_mem;
-    func_mem.load_elf_file( valid_elf_file);
+    ::load_elf_file( &func_mem, valid_elf_file);
 
     // read 4 bytes from the func_mem start addr
     uint64 right_ret = 0x03020100;
@@ -79,7 +87,7 @@ TEST_CASE( "Func_memory: Read_Method_Test")
 TEST_CASE( "Func_memory: Write_Read_Initialized_Mem_Test")
 {
     FuncMemory func_mem;
-    func_mem.load_elf_file( valid_elf_file);
+    ::load_elf_file( &func_mem, valid_elf_file);
 
     // write 1 into the byte pointed by dataSectAddr
     func_mem.write<uint8>( 1, dataSectAddr);
@@ -100,7 +108,7 @@ TEST_CASE( "Func_memory: Write_Read_Initialized_Mem_Test")
 TEST_CASE( "Func_memory: Write_Read_Not_Initialized_Mem_Test")
 {
     FuncMemory func_mem;
-    func_mem.load_elf_file( valid_elf_file);
+    ::load_elf_file( &func_mem, valid_elf_file);
 
     uint64 write_addr = 0x3FFFFE;
 
@@ -119,7 +127,7 @@ TEST_CASE( "Func_memory: Write_Read_Not_Initialized_Mem_Test")
 TEST_CASE( "Func_memory: Dump")
 {
     FuncMemory func_mem;
-    func_mem.load_elf_file( valid_elf_file);
+    ::load_elf_file( &func_mem, valid_elf_file);
 
     CHECK( func_mem.dump() ==
         "addr 0x400095: data 0xc\n"
@@ -154,8 +162,8 @@ TEST_CASE( "Func_memory: Invariancy")
 {
     FuncMemory mem1;
     FuncMemory mem2( 48, 15, 10);
-    mem1.load_elf_file( valid_elf_file);
-    mem2.load_elf_file( valid_elf_file);
+    ::load_elf_file( &mem1, valid_elf_file);
+    ::load_elf_file( &mem2, valid_elf_file);
     
     CHECK( mem1.dump() == mem2.dump());
